src/requirements.txt
astroid==2.5.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
idna==2.10
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.6.0
lxml==4.6.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.20.2
pandas==1.2.3
Pillow==8.2.0
pycodestyle==2.7.0
pydocstyle==6.0.0
pylint==2.7.4
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2021.1
requests==2.25.1
six==1.15.0
snowballstemmer==2.1.0
toml==0.10.2
typed-ast==1.4.2
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
